refactor: extract power-jobs.ts from daemon/main.ts
Move the 4 power-managed job registrations (embedding-reconcile,
session-maintenance, log-retention, auto-backup) into a dedicated
module with a single registerPowerJobs() entry point.
